Capa de negocio

+++ DTO/src/main/java/dto/CiudadanoDTO.java
@@ -10,13 +10,13 @@ public class CiudadanoDTO {
     private String apellidoP;
     private String apellidoM;
     private String CURP;
-    private int telefono;
+    private Long telefono;
     private String correo;
 
     public CiudadanoDTO() {
     }
 
-    public CiudadanoDTO(String nombre, String apellidoP, String apellidoM, String CURP, int telefono, String correo) {
+    public CiudadanoDTO(String nombre, String apellidoP, String apellidoM, String CURP, Long telefono, String correo) {
         this.nombre = nombre;
         this.apellidoP = apellidoP;
         this.apellidoM = apellidoM;
@@ -57,11 +57,11 @@ public class CiudadanoDTO {
         this.CURP = CURP;
     }
 
-    public int getTelefono() {
+    public Long getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(Long telefono) {
         this.telefono = telefono;
     }
